tests/wpt/tests/digital-credentials/digital-credentials-static-methods.tentative.https.html

<!DOCTYPE html>
<title>Digital Credential static methods.</title>
<link rel="help" href="https://wicg.github.io/digital-credentials/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body></body>
<script type="module">
  /**
   * @see https://www.w3.org/TR/credential-management/#dom-credential-isconditionalmediationavailable
   */
  promise_test(async (t) => {
    const result = await DigitalCredential.isConditionalMediationAvailable();
    assert_equals(result, false);
  }, "check default isConditionalMediationAvailable() behavior, inherited from Credential interface.");

  /**
   * @see https://www.w3.org/TR/credential-management/#dom-credential-willrequestconditionalcreation
   */
  promise_test(async (t) => {
    const result = await DigitalCredential.willRequestConditionalCreation();
    assert_equals(result, undefined);
  }, "check default willRequestConditionalCreation() behavior, inherited from Credential interface");
</script>
